float64 -> float32

+++ aircon/daikin/daikin01/entry.go
@@ -35,7 +35,7 @@ func (r *daikin01) Generate(e *aircon.Entry) ([]*hex.HexCode, error) {
 	}
 
 	// Temp
-	temp := e.Temp.(float64)
+	temp := e.Temp.(float32)
 	if e.Mode == "cool" || e.Mode == "heat" {
 		code[3][6] = 0x24 + int((temp-18.0)*2.0)
 	} else if e.Mode == "dry" || e.Mode == "auto" {
