+++ aircon/daikin/daikin03/entry.go
@@ -38,7 +38,7 @@ func (r *daikin03) Generate(e *aircon.Entry) ([]*hex.HexCode, error) {
 	}
 
 	// Temp
-	temp := e.Temp.(float64)
+	temp := e.Temp.(float32)
 	if e.Mode == "cool" || e.Mode == "heat" {
 		code[2][6] = 0x24 + int((temp-18.0)*2.0)
 	} else if e.Mode == "dry" || e.Mode == "auto" {
